fix: gio trash handle long file name fail
When gio deletes a file, it will create an info file for the deleted
file. The name of the info file is based on the original name with
the suffix ".trashinfo" added and uniquely processed. When the name
of the deleted file is too long, it will fail to create the info file,
or fail to write data to the info file.
The solution is to reduce the length of the info file name.

Log: fix gio trash bug
Bug: https://pms.uniontech.com/bug-view-224677.html